feat: repair lint:fix and refactor api folder
