+++ src/api/index.ts
@@ -1,0 +1,6 @@
+export { createUser } from './createUser';
+export { createWord } from './createWord';
+export { getAllWords } from './getAllWords';
+export { getUser } from './getUser';
+export { getWord } from './getWord';
+export { loginUser } from './signIn';
